move impl<Extensions..> from synth.rs to model.rs
